Move Config class to own file
